причесывание кода

+++ light.h
@@ -18,9 +18,9 @@ protected:
 
    GLfloat m_xRotate;
    GLfloat m_yRotate;
-   GLdouble m_xGo;
-   GLdouble m_yGo;
-   GLdouble m_Rotate;
+   GLfloat m_xGo;
+   GLfloat m_yGo;
+   GLfloat m_Rotate;
    QPoint m_ptPosition;
    void drawAxis();
    virtual void initializeGL();
